Move Integration type to its own file to make imports less cyclic
