+++ resources/ts/Components/Integrations/Detail/DeleteIntegration.tsx
@@ -5,8 +5,8 @@ import { faTrash } from "@fortawesome/free-solid-svg-icons";
 import { QuestionDialog } from "../../QuestionDialog";
 import { router } from "@inertiajs/react";
 import { t } from "i18next";
-import type { Integration } from "../../../Pages/Integrations/Index";
 import { Heading } from "../../Heading";
+import { Integration } from "../../../types/Integration";
 
 type Props = Integration;
