+++ resources/ts/Components/Integrations/Detail/IntegrationSettings.tsx
@@ -2,7 +2,6 @@ import React, { useEffect, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { ButtonPrimary } from "../../ButtonPrimary";
 import { useForm } from "@inertiajs/react";
-import type { Integration } from "../../../Pages/Integrations/Index";
 import { IntegrationUrlType } from "../../../types/IntegrationUrlType";
 import { UrlList } from "./UrlList";
 import { BasicInfo } from "./BasicInfo";
@@ -12,6 +11,7 @@ import { Environment } from "../../../types/Environment";
 import { QuestionDialog } from "../../QuestionDialog";
 import { KeyVisibility } from "../../../types/KeyVisibility";
 import type { IntegrationUrl } from "../../../types/IntegrationUrl";
+import { Integration } from "../../../types/Integration";
 
 export const NEW_URL_ID_PREFIX = "new-";
